requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pdf2image
pyjwt
